[REFACTOR/#1528] replace stamp NetworkErrorDialog with NetworkDialog
- Remove local `ErrorDialog` and `NetworkErrorDialog` implementations in `feature/soptamp`.
- Migrate to `org.sopt.official.designsystem.component.dialog.NetworkErrorDialog`.
- Update `onRetry` callback parameter to `onConfirm` across mission and ranking screens.
